% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{data.prep}
\alias{data.prep}
\title{Data reading from raw data}
\usage{
data.prep(code)
}
\arguments{
\item{code}{}
}
\value{
Data matrix with fitting structure to be used in \code{fit_TIM_2P()}.
}
\description{
Data reading from raw data
}
